# src/swagger.yaml
swagger: '2.0'
info:
  description: Task Manager API
  title: Task Manager API
  version: 0.0.1
host: 'localhost:10001'
schemes:
  - https
  - http
basePath: /api/v0
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: implicit
    scopes:
      'write:pets': modify pets in your account
      'read:pets': read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
  apiKeyAuth:
    type: apiKey
    name: X-Auth-Key
    in: header
...
paths:
  /task:
    post:
      tags:
        - task
      summary: Add a new task
      description: ''
      operationId: addTask
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - in: body
          name: body
          description: Task object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/Task'
      responses:
        '405':
          description: Invalid input
      x-swagger-router-controller: Task
    get:
      tags:
        - task
      summary: Get task list
      description: ''
      operationId: getTaskList
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - name: page
          in: query
          type: integer
        - name: pagesize
          in: query
          type: integer
        - name: next
          in: query
          type: integer
        - name: filters
          in: query
          type: string
        - name: orderBy
          in: query
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            properties:
              total:
                type: integer
                format: int32
              items:
                type: array
                items:
                  $ref: '#/definitions/Task'
        '400':
          description: Get task list failed
      x-swagger-router-controller: Task
  '/task/{taskId}':
    get:
      tags:
        - task
      summary: Find task by ID
      description: Returns a single task
      operationId: getTask
      produces:
        - application/json
        - application/xml
      parameters:
        - name: taskId
          in: path
          description: ID of task to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Task'
        '400':
          description: Invalid ID supplied
        '404':
          description: Task not found
      x-swagger-router-controller: Task
    post:
      tags:
        - task
      summary: Updates a task with form data
      description: ''
      operationId: updateTask
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - name: taskId
          in: path
          description: ID of task that needs to be updated
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description: Task object that needs to be updated
          required: true
          schema:
            $ref: '#/definitions/Task'
      responses:
        '200':
          description: successful operation
        '405':
          description: Invalid input
      x-swagger-router-controller: Task
    delete:
      tags:
        - task
      summary: Deletes a task
      description: ''
      operationId: deleteTask
      produces:
        - application/xml
        - application/json
      parameters:
        - name: taskId
          in: path
          description: Task id to delete
          required: true
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Task not found
      x-swagger-router-controller: Task
...
definitions:
  Task:
    type: object
    required:
      - title
    properties:
      id:
        type: integer
        format: int64
      isCompleted:
        type: boolean
        default: false
      title:
        type: string
      content:
        type: string
      deadline:
        type: integer
        format: int64
    xml:
      name: Task
    example:
      id: 0
      isCompleted: false
      title: title
      content: content
      deadline: 0
...